// frontend/web/src/services/API/proserve-wb-packaging-api/models/GetComponentsVersionsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ComponentVersionSummary } from './ComponentVersionSummary';
import {
    ComponentVersionSummaryFromJSON,
    ComponentVersionSummaryFromJSONTyped,
    ComponentVersionSummaryToJSON,
} from './ComponentVersionSummary';

/**
 * 
 * @export
 * @interface GetComponentsVersionsResponse
 */
export interface GetComponentsVersionsResponse {
    /**
     * 
     * @type {Array<ComponentVersionSummary>}
     * @memberof GetComponentsVersionsResponse
     */
    componentsVersionsSummary: Array<ComponentVersionSummary>;
}

/**
 * Check if a given object implements the GetComponentsVersionsResponse interface.
 */
export function instanceOfGetComponentsVersionsResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "componentsVersionsSummary" in value;

    return isInstance;
}

export function GetComponentsVersionsResponseFromJSON(json: any): GetComponentsVersionsResponse {
    return GetComponentsVersionsResponseFromJSONTyped(json, false);
}

export function GetComponentsVersionsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetComponentsVersionsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'componentsVersionsSummary': ((json['components_versions_summary'] as Array<any>).map(ComponentVersionSummaryFromJSON)),
    };
}

export function GetComponentsVersionsResponseToJSON(value?: GetComponentsVersionsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'components_versions_summary': ((value.componentsVersionsSummary as Array<any>).map(ComponentVersionSummaryToJSON)),
    };
}