// frontend/web/src/services/API/proserve-wb-packaging-api/models/ComponentVersionSummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ComponentVersionSummary
 */
export interface ComponentVersionSummary {
    /**
     * Component ID
     * @type {string}
     * @memberof ComponentVersionSummary
     */
    componentId: string;
    /**
     * Component version ID
     * @type {string}
     * @memberof ComponentVersionSummary
     */
    componentVersionId: string;
    /**
     * Component version Name
     * @type {string}
     * @memberof ComponentVersionSummary
     */
    componentVersionName: string;
    /**
     * Component Name
     * @type {string}
     * @memberof ComponentVersionSummary
     */
    componentName: string;
}

/**
 * Check if a given object implements the ComponentVersionSummary interface.
 */
export function instanceOfComponentVersionSummary(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "componentId" in value;
    isInstance = isInstance && "componentVersionId" in value;
    isInstance = isInstance && "componentVersionName" in value;
    isInstance = isInstance && "componentName" in value;

    return isInstance;
}

export function ComponentVersionSummaryFromJSON(json: any): ComponentVersionSummary {
    return ComponentVersionSummaryFromJSONTyped(json, false);
}

export function ComponentVersionSummaryFromJSONTyped(json: any, ignoreDiscriminator: boolean): ComponentVersionSummary {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'componentId': json['componentId'],
        'componentVersionId': json['componentVersionId'],
        'componentVersionName': json['componentVersionName'],
        'componentName': json['componentName'],
    };
}

export function ComponentVersionSummaryToJSON(value?: ComponentVersionSummary | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'componentId': value.componentId,
        'componentVersionId': value.componentVersionId,
        'componentVersionName': value.componentVersionName,
        'componentName': value.componentName,
    };
}